Upgrade Mario's jumping (Mario jumps different heights based on Mario's horizontal speed)

+++ DoAn-SE102/Mushroom.cpp
@@ -12,7 +12,7 @@ void CMushroom::GetBoundingBox(float& left, float& top, float& right, float& bot
 
 void CMushroom::Render() {
 	float cX, cY; CGame::GetInstance()->GetCamPos(cX, cY);
-	float screenWidth = CGame::GetInstance()->GetBackBufferWidth(), screenHeight = CGame::GetInstance()->GetBackBufferHeight();
+	int screenWidth = CGame::GetInstance()->GetBackBufferWidth(), screenHeight = CGame::GetInstance()->GetBackBufferHeight();
 	if (x < cX - MUSHROOM_WIDTH * 0.5f || x > cX + screenWidth + MUSHROOM_WIDTH * 0.5f) return;
 	if (y < cY - MUSHROOM_HEIGHT * 0.5f || y > cY + screenHeight + MUSHROOM_HEIGHT * 0.5f) { isDeleted = true; return; }
 	CSprites::GetInstance()->Get(MUSHROOM_SPRITE)->Draw(x, y);
